Fixed task list - now showing correct items

+++ androidApp/src/main/java/co/softov/morestuff/androidApp/domain/Debug.kt
@@ -3,6 +3,6 @@ package co.softov.morestuff.androidApp.domain
 interface Debug {
 
     val debugReminders: Boolean
-    val todayDebugTime: Long
+    val todayDebugTime: Int
 
 }
